Move collated string array operation to CollationExpressionSuite and reduce checks in collations.sql
