+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/CollectionExpressionsSuite.scala
@@ -29,7 +29,7 @@ import org.apache.spark.sql.Row
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult.DataTypeMismatch
-import org.apache.spark.sql.catalyst.util.{CollationFactory, DateTimeTestUtils, DateTimeUtils}
+import org.apache.spark.sql.catalyst.util.{DateTimeTestUtils, DateTimeUtils}
 import org.apache.spark.sql.catalyst.util.DateTimeTestUtils.{outstandingZoneIds, LA, UTC}
 import org.apache.spark.sql.catalyst.util.IntervalUtils._
 import org.apache.spark.sql.catalyst.util.TypeUtils.ordinalNumber
@@ -2820,89 +2820,5 @@ class CollectionExpressionsSuite
     checkEvaluation(new ArrayInsert(
       a, Literal(5), Literal.create("q", StringType)), Seq("b", "a", "c", null, "q")
     )
-  }
-
-  test("Array operations on arrays of collated strings") {
-    val arrayLiteral = (arr: Seq[String], collName: String) =>
-      Literal.create(
-        arr,
-        ArrayType(StringType(CollationFactory.collationNameToId(collName)))
-      )
-
-    // arrays_overlap
-    val overlap = Seq(
-      (Seq("a"), Seq("a"), true, "UTF8_BINARY"),
-      (Seq("a"), Seq("b"), false, "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), false, "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), true, "UTF8_BINARY_LCASE"),
-      (Seq("a", "B"), Seq("A", "b"), true, "UTF8_BINARY_LCASE"),
-      (Seq("a"), Seq("A"), false, "UNICODE"),
-      (Seq("a", "B"), Seq("A", "b"), true, "UNICODE_CI")
-    )
-    for ((inLeft, inRight, out, collName) <- overlap) {
-      val left = arrayLiteral(inLeft, collName)
-      val right = arrayLiteral(inRight, collName)
-      checkEvaluation(ArraysOverlap(left, right), out)
-    }
-
-    // array_distinct
-    val distinct = Seq(
-      (Seq("a", "b", "c"), Seq("a", "b", "c"), "UTF8_BINARY"),
-      (Seq("a", "a", "a"), Seq("a"), "UTF8_BINARY"),
-      (Seq("aaa", "AAA", "Aaa", "aAa"), Seq("aaa", "AAA", "Aaa", "aAa"), "UTF8_BINARY"),
-      (Seq("aaa", "AAA", "Aaa", "aAa"), Seq("aaa"), "UTF8_BINARY_LCASE"),
-      (Seq("aaa", "AAA", "Aaa", "aAa", "b"), Seq("aaa", "b"), "UTF8_BINARY_LCASE"),
-      (Seq("aaa", "AAA", "Aaa", "aAa"), Seq("aaa"), "UNICODE_CI")
-    )
-    for ((in, out, collName) <- distinct)
-      checkEvaluation(ArrayDistinct(arrayLiteral(in, collName)), out)
-
-    // array_union
-    val union = Seq(
-      (Seq("a"), Seq("a"), Seq("a"), "UTF8_BINARY"),
-      (Seq("a"), Seq("b"), Seq("a", "b"), "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), Seq("a", "A"), "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), Seq("a"), "UTF8_BINARY_LCASE"),
-      (Seq("a", "B"), Seq("A", "b"), Seq("a", "B"), "UTF8_BINARY_LCASE"),
-      (Seq("a"), Seq("A"), Seq("a", "A"), "UNICODE"),
-      (Seq("a", "B"), Seq("A", "b"), Seq("a", "B"), "UNICODE_CI")
-    )
-    for ((inLeft, inRight, out, collName) <- union) {
-      val left = arrayLiteral(inLeft, collName)
-      val right = arrayLiteral(inRight, collName)
-      checkEvaluation(ArrayUnion(left, right), out)
-    }
-
-    // array_intersect
-    val intersect = Seq(
-      (Seq("a"), Seq("a"), Seq("a"), "UTF8_BINARY"),
-      (Seq("a"), Seq("b"), Seq.empty, "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), Seq.empty, "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), Seq("a"), "UTF8_BINARY_LCASE"),
-      (Seq("a", "B"), Seq("A", "b"), Seq("a", "B"), "UTF8_BINARY_LCASE"),
-      (Seq("a"), Seq("A"), Seq.empty, "UNICODE"),
-      (Seq("a", "B"), Seq("A", "b"), Seq("a", "B"), "UNICODE_CI")
-    )
-    for ((inLeft, inRight, out, collName) <- intersect) {
-      val left = arrayLiteral(inLeft, collName)
-      val right = arrayLiteral(inRight, collName)
-      checkEvaluation(ArrayIntersect(left, right), out)
-    }
-
-    // array_except
-    val except = Seq(
-      (Seq("a"), Seq("a"), Seq.empty, "UTF8_BINARY"),
-      (Seq("a"), Seq("b"), Seq("a"), "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), Seq("a"), "UTF8_BINARY"),
-      (Seq("a"), Seq("A"), Seq.empty, "UTF8_BINARY_LCASE"),
-      (Seq("a", "B"), Seq("A", "b"), Seq.empty, "UTF8_BINARY_LCASE"),
-      (Seq("a"), Seq("A"), Seq("a"), "UNICODE"),
-      (Seq("a", "B"), Seq("A", "b"), Seq.empty, "UNICODE_CI")
-    )
-    for ((inLeft, inRight, out, collName) <- except) {
-      val left = arrayLiteral(inLeft, collName)
-      val right = arrayLiteral(inRight, collName)
-      checkEvaluation(ArrayExcept(left, right), out)
-    }
   }
 }
