+++ sql/core/src/test/resources/sql-tests/results/collations.sql.out
@@ -310,27 +310,11 @@ true
 
 
 -- !query
-select arrays_overlap(array('aaa' collate utf8_binary_lcase), array('AAAB' collate utf8_binary_lcase))
--- !query schema
-struct<arrays_overlap(array(collate(aaa)), array(collate(AAAB))):boolean>
--- !query output
-false
-
-
--- !query
 select array_distinct(array('aaa' collate utf8_binary_lcase, 'AAA' collate utf8_binary_lcase))
 -- !query schema
 struct<array_distinct(array(collate(aaa), collate(AAA))):array<string collate UTF8_BINARY_LCASE>>
 -- !query output
 ["aaa"]
-
-
--- !query
-select array_distinct(array('aaa' collate utf8_binary_lcase, 'AAAB' collate utf8_binary_lcase))
--- !query schema
-struct<array_distinct(array(collate(aaa), collate(AAAB))):array<string collate UTF8_BINARY_LCASE>>
--- !query output
-["aaa","AAAB"]
 
 
 -- !query
@@ -342,14 +326,6 @@ struct<array_union(array(collate(aaa)), array(collate(AAA))):array<string collat
 
 
 -- !query
-select array_union(array('aaa' collate utf8_binary_lcase), array('AAAB' collate utf8_binary_lcase))
--- !query schema
-struct<array_union(array(collate(aaa)), array(collate(AAAB))):array<string collate UTF8_BINARY_LCASE>>
--- !query output
-["aaa","AAAB"]
-
-
--- !query
 select array_intersect(array('aaa' collate utf8_binary_lcase), array('AAA' collate utf8_binary_lcase))
 -- !query schema
 struct<array_intersect(array(collate(aaa)), array(collate(AAA))):array<string collate UTF8_BINARY_LCASE>>
@@ -358,24 +334,8 @@ struct<array_intersect(array(collate(aaa)), array(collate(AAA))):array<string co
 
 
 -- !query
-select array_intersect(array('aaa' collate utf8_binary_lcase), array('AAAB' collate utf8_binary_lcase))
--- !query schema
-struct<array_intersect(array(collate(aaa)), array(collate(AAAB))):array<string collate UTF8_BINARY_LCASE>>
--- !query output
-[]
-
-
--- !query
 select array_except(array('aaa' collate utf8_binary_lcase), array('AAA' collate utf8_binary_lcase))
 -- !query schema
 struct<array_except(array(collate(aaa)), array(collate(AAA))):array<string collate UTF8_BINARY_LCASE>>
 -- !query output
 []
-
-
--- !query
-select array_except(array('aaa' collate utf8_binary_lcase), array('AAAB' collate utf8_binary_lcase))
--- !query schema
-struct<array_except(array(collate(aaa)), array(collate(AAAB))):array<string collate UTF8_BINARY_LCASE>>
--- !query output
-["aaa"]
